+++ src/components/registration/Registration.utils.ts
@@ -1,3 +1,4 @@
+import { UtmParameters } from '@/src/hooks/queryParams/useUTM';
 import { RegistrableUserRole } from 'src/constants/users';
 import {
   FlattenedRegistrationFormData,
@@ -30,4 +31,17 @@ export const incrementRegistrationStep = (
 ): RegistrationStep => {
   const stepNumber: number = parseInt(step.split('-')[1], 10);
   return `step-${stepNumber + 1}`;
+};
+
+export const getUtmFromLocalStorage = (): UtmParameters[] => {
+  const utmParameters: UtmParameters[] = [];
+
+  Object.values(UtmParameters).forEach((key) => {
+    const value = localStorage.getItem(key);
+    if (value) {
+      utmParameters[key] = value;
+    }
+  });
+
+  return utmParameters;
 };
